requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy<2
torch==2.1.0
transformers==4.35.2
tokenizers==0.15.0
sentencepiece==0.1.99
sentence-transformers==2.2.2
huggingface-hub==0.19.4
safetensors==0.4.1
faiss-cpu==1.7.4
pandas==2.0.3
scikit-learn==1.3.2
tiktoken==0.5.2
python-dotenv==1.0.0
streamlit==1.29.0
protobuf==4.25.1
elasticsearch==8.11.0
langchain==0.1.0
langchain-community==0.0.10
pymongo==4.6.0
psycopg2-binary==2.9.9
mlflow==2.21.0